+++ clang/docs/ReleaseNotes.rst
@@ -192,6 +192,7 @@ C++ Language Changes
 
 - ``__is_trivially_equality_comparable`` no longer returns false for all enum types. (#GH132672)
 - ``auto`` parameters are now available in all C++ language modes as an extension.
+- Clang now supports friend declarations with a dependent nested name specifier. (#GH104057)
 
 C++2c Feature Support
 ^^^^^^^^^^^^^^^^^^^^^